+++ lib/adaptive_media_picker.dart
@@ -1,3 +1,5 @@
-export 'src/models.dart';
+export 'src/core/models.dart';
 export 'src/adaptive_media_picker_impl.dart' show AdaptiveMediaPicker;
 export 'package:image_picker/image_picker.dart' show ImageSource;
+export 'src/platform/permission_manager.dart' show PermissionManager;
+export 'src/ui/limited_access_picker.dart' show LimitedAccessPicker;